(PC-17816) refactor(profile): remove __tests__ folder for DeleteProfile
