+++ src/features/profile/pages/DeleteProfile/DeleteProfileSuccess.native.test.tsx
@@ -7,7 +7,7 @@ import { navigateFromRef } from 'features/navigation/navigationRef'
 import { analytics } from 'libs/firebase/analytics'
 import { fireEvent, render } from 'tests/utils'
 
-import { DeleteProfileSuccess } from '../DeleteProfileSuccess'
+import { DeleteProfileSuccess } from './DeleteProfileSuccess'
 
 jest.mock('features/navigation/helpers')
 jest.mock('features/navigation/navigationRef')
